# src/main/java/com/codeit/demo/api-docs.json
{"openapi":"3.1.0","info":{"title":"HR Bank API","description":"HR Bank API 문서","version":"v1.0"},"servers":[{"url":"http://localhost:8080","description":"Local Server"}],"tags":[{"name":"직원 관리","description":"직원 관리 API"},{"name":"부서 관리","description":"부서 관리 API"},{"name":"데이터 백업 관리","description":"데이터 백업 관리 API"},{"name":"직원 정보 수정 이력 관리","description":"직원 정보 수정 이력 관리 API"},{"name":"파일 관리","description":"파일 관리 API"}],"paths":{"/api/employees":{"get":{"tags":["직원 관리"],"summary":"직원 목록 조회","description":"직원 목록을 조회합니다.","operationId":"getAllEmployees","parameters":[{"name":"nameOrEmail","in":"query","description":"직원 이름 또는 이메일","required":false,"schema":{"type":"string"}},{"name":"employeeNumber","in":"query","description":"사원 번호","required":false,"schema":{"type":"string"}},{"name":"departmentName","in":"query","description":"부서 이름","required":false,"schema":{"type":"string"}},{"name":"position","in":"query","description":"직함","required":false,"schema":{"type":"string"}},{"name":"hireDateFrom","in":"query","description":"입사일 시작","required":false,"schema":{"type":"string","format":"date"}},{"name":"hireDateTo","in":"query","description":"입사일 종료","required":false,"schema":{"type":"string","format":"date"}},{"name":"status","in":"query","description":"상태 (재직중, 휴직중, 퇴사)","required":false,"schema":{"type":"string","enum":["ACTIVE","ON_LEAVE","RESIGNED"]}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (다음 페이지 시작점)","required":false,"schema":{"type":"string"}},{"name":"size","in":"query","description":"페이지 크기 (기본값: 10)","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortField","in":"query","description":"정렬 필드 (name, employeeNumber, hireDate)","required":false,"schema":{"type":"string","default":"name"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc 또는 desc, 기본값: asc)","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseEmployeeDto"}}}}}},"post":{"tags":["직원 관리"],"summary":"직원 등록","description":"새로운 직원을 등록합니다.","operationId":"createEmployee","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"employee":{"$ref":"#/components/schemas/EmployeeCreateRequest"},"profile":{"type":"string","format":"binary","description":"프로필 이미지"}},"required":["employee"]}}}},"responses":{"200":{"description":"등록 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"부서를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 중복된 이메일","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/departments":{"get":{"tags":["부서 관리"],"summary":"부서 목록 조회","description":"부서 목록을 조회합니다.","operationId":"getAllDepartments","parameters":[{"name":"nameOrDescription","in":"query","description":"부서 이름 또는 설명","required":false,"schema":{"type":"string"}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (다음 페이지 시작점)","required":false,"schema":{"type":"string"}},{"name":"size","in":"query","description":"페이지 크기 (기본값: 10)","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortField","in":"query","description":"정렬 필드 (name 또는 establishedDate)","required":false,"schema":{"type":"string","default":"establishedDate"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc 또는 desc, 기본값: asc)","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseDepartmentDto"}}}}}},"post":{"tags":["부서 관리"],"summary":"부서 등록","description":"새로운 부서를 등록합니다.","operationId":"createDepartment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentCreateRequest"}}},"required":true},"responses":{"200":{"description":"등록 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DepartmentDto"}}}},"400":{"description":"잘못된 요청 또는 중복된 이름","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/backups":{"get":{"tags":["데이터 백업 관리"],"summary":"데이터 백업 목록 조회","description":"데이터 백업 목록을 조회합니다.","operationId":"getAllBackups","parameters":[{"name":"worker","in":"query","description":"작업자","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","description":"상태 (IN_PROGRESS, COMPLETED, FAILED)","required":false,"schema":{"type":"string"}},{"name":"startedAtFrom","in":"query","description":"시작 시간(부터)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"startedAtTo","in":"query","description":"시작 시간(까지)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (이전 페이지의 마지막 ID)","required":false,"schema":{"type":"string"}},{"name":"size","in":"query","description":"페이지 크기","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortField","in":"query","description":"정렬 필드 (startedAt, endedAt, status)","required":false,"schema":{"type":"string","default":"startedAt"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (ASC, DESC)","required":false,"schema":{"type":"string","default":"DESC"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 지원하지 않는 정렬 필드","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseBackupDto"}}}}}},"post":{"tags":["데이터 백업 관리"],"summary":"데이터 백업 생성","description":"데이터 백업을 생성합니다.","operationId":"createBackup","responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"백업 생성 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BackupDto"}}}},"400":{"description":"잘못된 요청","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"이미 진행 중인 백업이 있음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/employees/{id}":{"get":{"tags":["직원 관리"],"summary":"직원 상세 조회","description":"직원 상세 정보를 조회합니다.","operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","description":"직원 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"직원을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}}}},"delete":{"tags":["직원 관리"],"summary":"직원 삭제","description":"직원을 삭제합니다.","operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","description":"직원 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"직원을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"204":{"description":"삭제 성공"}}},"patch":{"tags":["직원 관리"],"summary":"직원 수정","description":"직원 정보를 수정합니다.","operationId":"updateEmployee","parameters":[{"name":"id","in":"path","description":"직원 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"employee":{"$ref":"#/components/schemas/EmployeeUpdateRequest"},"profile":{"type":"string","format":"binary","description":"프로필 이미지"}},"required":["employee"]}}}},"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}},"404":{"description":"직원 또는 부서를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 중복된 이메일","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/departments/{id}":{"get":{"tags":["부서 관리"],"summary":"부서 상세 조회","description":"부서 상세 정보를 조회합니다.","operationId":"getDepartmentById","parameters":[{"name":"id","in":"path","description":"부서 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"부서를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DepartmentDto"}}}}}},"delete":{"tags":["부서 관리"],"summary":"부서 삭제","description":"부서를 삭제합니다.","operationId":"deleteDepartment","parameters":[{"name":"id","in":"path","description":"부서 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"소속 직원이 있는 부서는 삭제할 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"부서를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"204":{"description":"삭제 성공"}}},"patch":{"tags":["부서 관리"],"summary":"부서 수정","description":"부서 정보를 수정합니다.","operationId":"updateDepartment","parameters":[{"name":"id","in":"path","description":"부서 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentUpdateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청 또는 중복된 이름","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DepartmentDto"}}}},"404":{"description":"부서를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/files/{id}/download":{"get":{"tags":["파일 관리"],"summary":"파일 다운로드","description":"파일을 다운로드합니다.","operationId":"downloadFile","parameters":[{"name":"id","in":"path","description":"파일 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"다운로드 성공","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"파일을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/employees/stats/trend":{"get":{"tags":["직원 관리"],"summary":"직원 수 추이 조회","description":"지정된 기간 및 시간 단위로 그룹화된 직원 수 추이를 조회합니다. 파라미터를 제공하지 않으면 최근 12개월 데이터를 월 단위로 반환합니다.","operationId":"getTrend","parameters":[{"name":"from","in":"query","description":"시작 일시 (기본값: 현재로부터 unit 기준 12개 이전)","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","description":"종료 일시 (기본값: 현재)","required":false,"schema":{"type":"string","format":"date"}},{"name":"unit","in":"query","description":"시간 단위 (day, week, month, quarter, year, 기본값: month)","required":false,"schema":{"type":"string","default":"month"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 지원하지 않는 시간 단위","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeTrendDto"}}}}}}}},"/api/employees/stats/distribution":{"get":{"tags":["직원 관리"],"summary":"직원 분포 조회","description":"지정된 기준으로 그룹화된 직원 분포를 조회합니다.","operationId":"getDistribution","parameters":[{"name":"groupBy","in":"query","description":"그룹화 기준 (department: 부서별, position: 직무별, 기본값: department)","required":false,"schema":{"type":"string","default":"department"}},{"name":"status","in":"query","description":"직원 상태 (재직중, 휴직중, 퇴사, 기본값: 재직중)","required":false,"schema":{"type":"string","default":"ACTIVE","enum":["ACTIVE","ON_LEAVE","RESIGNED"]}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 지원하지 않는 그룹화 기준","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeDistributionDto"}}}}}}}},"/api/employees/count":{"get":{"tags":["직원 관리"],"summary":"직원 수 조회","description":"지정된 조건에 맞는 직원 수를 조회합니다. 상태 필터링 및 입사일 기간 필터링이 가능합니다.","operationId":"getEmployeeCount","parameters":[{"name":"status","in":"query","description":"직원 상태 (재직중, 휴직중, 퇴사)","required":false,"schema":{"type":"string","enum":["ACTIVE","ON_LEAVE","RESIGNED"]}},{"name":"fromDate","in":"query","description":"입사일 시작 (지정 시 해당 기간 내 입사한 직원 수 조회, 미지정 시 전체 직원 수 조회)","required":false,"schema":{"type":"string","format":"date"}},{"name":"toDate","in":"query","description":"입사일 종료 (fromDate와 함께 사용, 기본값: 현재 일시)","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/api/change-logs":{"get":{"tags":["직원 정보 수정 이력 관리"],"summary":"직원 정보 수정 이력 목록 조회","description":"직원 정보 수정 이력 목록을 조회합니다. 상세 변경 내용은 포함되지 않습니다.","operationId":"getAllChangeLogs","parameters":[{"name":"employeeNumber","in":"query","description":"대상 직원 사번","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","description":"이력 유형 (CREATED, UPDATED, DELETED)","required":false,"schema":{"type":"string"}},{"name":"memo","in":"query","description":"내용","required":false,"schema":{"type":"string"}},{"name":"ipAddress","in":"query","description":"IP 주소","required":false,"schema":{"type":"string"}},{"name":"atFrom","in":"query","description":"수정 일시(부터)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"atTo","in":"query","description":"수정 일시(까지)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (이전 페이지의 마지막 ID)","required":false,"schema":{"type":"string"}},{"name":"size","in":"query","description":"페이지 크기","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortField","in":"query","description":"정렬 필드 (ipAddress, at)","required":false,"schema":{"type":"string","default":"at"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc, desc)","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 지원하지 않는 정렬 필드","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseChangeLogDto"}}}}}}},"/api/change-logs/{id}/diffs":{"get":{"tags":["직원 정보 수정 이력 관리"],"summary":"직원 정보 수정 이력 상세 조회","description":"직원 정보 수정 이력의 상세 정보를 조회합니다. 변경 상세 내용이 포함됩니다.","operationId":"getChangeLogDiffs","parameters":[{"name":"id","in":"path","description":"이력 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"이력을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DiffDto"}}}}}}}},"/api/change-logs/count":{"get":{"tags":["직원 정보 수정 이력 관리"],"summary":"수정 이력 건수 조회","description":"직원 정보 수정 이력 건수를 조회합니다. 파라미터를 제공하지 않으면 최근 일주일 데이터를 반환합니다.","operationId":"getChangeLogsCount","parameters":[{"name":"fromDate","in":"query","description":"시작 일시 (기본값: 7일 전)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"toDate","in":"query","description":"종료 일시 (기본값: 현재)","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 유효하지 않은 날짜 범위","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/api/backups/latest":{"get":{"tags":["데이터 백업 관리"],"summary":"최근 백업 정보 조회","description":"지정된 상태의 가장 최근 백업 정보를 조회합니다. 상태를 지정하지 않으면 성공적으로 완료된(COMPLETED) 백업을 반환합니다.","operationId":"getLatestBackup","parameters":[{"name":"status","in":"query","description":"백업 상태 (COMPLETED, FAILED, IN_PROGRESS, 기본값: COMPLETED)","required":false,"schema":{"type":"string","default":"COMPLETED"}}],"responses":{"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 또는 유효하지 않은 상태값","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BackupDto"}}}}}}}},"components":{"schemas":{"EmployeeCreateRequest":{"type":"object","description":"직원 등록 요청","properties":{"name":{"type":"string","description":"직원 이름","example":"홍길동"},"email":{"type":"string","description":"직원 이메일","example":"hong@example.com"},"departmentId":{"type":"integer","format":"int64","description":"부서 ID","example":1},"position":{"type":"string","description":"직함","example":"선임 개발자"},"hireDate":{"type":"string","format":"date","description":"입사일","example":"2023-01-01"},"memo":{"type":"string","description":"수정 내용 (이력 기록용)","example":"신규 입사자 등록"}}},"ErrorResponse":{"type":"object","description":"에러 응답","properties":{"timestamp":{"type":"string","format":"date-time","description":"에러 발생 시간","example":"2025-03-06T05:39:06.152068Z"},"status":{"type":"integer","format":"int32","description":"HTTP 상태 코드","example":400},"message":{"type":"string","description":"에러 메시지","example":"잘못된 요청입니다."},"details":{"type":"string","description":"에러 상세 내용","example":"부서 코드는 필수입니다."}}},"EmployeeDto":{"type":"object","description":"직원 정보","properties":{"id":{"type":"integer","format":"int64","description":"직원 ID","example":1},"name":{"type":"string","description":"직원 이름","example":"홍길동"},"email":{"type":"string","description":"직원 이메일","example":"hong@example.com"},"employeeNumber":{"type":"string","description":"사원 번호","example":"EMP-2023-001"},"departmentId":{"type":"integer","format":"int64","description":"부서 ID","example":1},"departmentName":{"type":"string","description":"부서 이름","example":"개발팀"},"position":{"type":"string","description":"직함","example":"선임 개발자"},"hireDate":{"type":"string","format":"date","description":"입사일","example":"2023-01-01"},"status":{"type":"string","description":"상태 (재직중, 휴직중, 퇴사)","enum":["ACTIVE","ON_LEAVE","RESIGNED"],"example":"재직중"},"profileImageId":{"type":"integer","format":"int64","description":"프로필 이미지 ID","example":1}}},"DepartmentCreateRequest":{"type":"object","description":"부서 등록 요청","properties":{"name":{"type":"string","description":"부서 이름","example":"개발팀"},"description":{"type":"string","description":"부서 설명","example":"소프트웨어 개발을 담당하는 부서입니다."},"establishedDate":{"type":"string","format":"date","description":"부서 설립일","example":"2023-01-01"}}},"DepartmentDto":{"type":"object","description":"부서 정보","properties":{"id":{"type":"integer","format":"int64","description":"부서 ID","example":1},"name":{"type":"string","description":"부서 이름","example":"개발팀"},"description":{"type":"string","description":"부서 설명","example":"소프트웨어 개발을 담당하는 부서입니다."},"establishedDate":{"type":"string","format":"date","description":"부서 설립일","example":"2023-01-01"},"employeeCount":{"type":"integer","format":"int64","description":"소속 직원 수","example":10}}},"BackupDto":{"type":"object","description":"데이터 백업 정보 (목록 조회용)","properties":{"id":{"type":"integer","format":"int64","description":"백업 ID","example":1},"worker":{"type":"string","description":"작업자","example":"192.168.0.1"},"startedAt":{"type":"string","format":"date-time","description":"시작 시간","example":"2023-01-01T12:00:00Z"},"endedAt":{"type":"string","format":"date-time","description":"종료 시간","example":"2023-01-01T12:05:00Z"},"status":{"type":"string","description":"상태 (진행중, 완료, 실패)","enum":["IN_PROGRESS","COMPLETED","SKIPPED","FAILED"],"example":"완료"},"fileId":{"type":"integer","format":"int64","description":"백업 파일 ID","example":1}}},"EmployeeUpdateRequest":{"type":"object","description":"직원 수정 요청","properties":{"name":{"type":"string","description":"직원 이름","example":"홍길동"},"email":{"type":"string","description":"직원 이메일","example":"hong@example.com"},"departmentId":{"type":"integer","format":"int64","description":"부서 ID","example":1},"position":{"type":"string","description":"직함","example":"선임 개발자"},"hireDate":{"type":"string","format":"date","description":"입사일","example":"2023-01-01"},"status":{"type":"string","description":"상태 (재직중, 휴직중, 퇴사)","enum":["ACTIVE","ON_LEAVE","RESIGNED"],"example":"재직중"},"memo":{"type":"string","description":"수정 내용 (이력 기록용)","example":"직함 변경"}}},"DepartmentUpdateRequest":{"type":"object","description":"부서 수정 요청","properties":{"name":{"type":"string","description":"부서 이름","example":"개발팀"},"description":{"type":"string","description":"부서 설명","example":"소프트웨어 개발을 담당하는 부서입니다."},"establishedDate":{"type":"string","format":"date","description":"부서 설립일","example":"2023-01-01"}}},"CursorPageResponseEmployeeDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/EmployeeDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":20},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}},"EmployeeTrendDto":{"type":"object","description":"직원 수 추이 정보","properties":{"date":{"type":"string","format":"date","description":"시점 (날짜)","example":"2023-01-01"},"count":{"type":"integer","format":"int64","description":"직원 수","example":150},"change":{"type":"integer","format":"int64","description":"증감 (이전 시점 대비)","example":5},"changeRate":{"type":"number","format":"double","description":"증감률 (이전 시점 대비, %)","example":3.5}}},"EmployeeDistributionDto":{"type":"object","description":"직원 분포 정보","properties":{"groupKey":{"type":"string","description":"분류명 (부서명 또는 직무명)","example":"개발팀"},"count":{"type":"integer","format":"int64","description":"직원 수","example":25},"percentage":{"type":"number","format":"double","description":"비율 (%)","example":16.7}}},"CursorPageResponseDepartmentDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/DepartmentDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":20},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}},"ChangeLogDto":{"type":"object","description":"직원 정보 수정 이력 (목록 조회용)","properties":{"id":{"type":"integer","format":"int64","description":"이력 ID","example":1},"type":{"type":"string","description":"유형 (직원 추가, 정보 수정, 직원 삭제)","enum":["CREATED","UPDATED","DELETED"],"example":"UPDATED"},"employeeNumber":{"type":"string","description":"직원 사번","example":"EMP-2023-001"},"memo":{"type":"string","description":"내용","example":"직함 변경에 따른 수정"},"ipAddress":{"type":"string","description":"IP 주소","example":"192.168.0.1"},"at":{"type":"string","format":"date-time","description":"수정 일시","example":"2023-01-01T12:00:00"}}},"CursorPageResponseChangeLogDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/ChangeLogDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":20},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}},"DiffDto":{"type":"object","description":"직원 정보 수정 이력 변경 내용 (상세 조회용)","properties":{"propertyName":{"type":"string","description":"속성 이름","example":"직함"},"before":{"type":"string","description":"수정 전 값","example":"사원"},"after":{"type":"string","description":"수정 후 값","example":"대리"}}},"CursorPageResponseBackupDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/BackupDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":20},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}}}}}